requirements.txt
langchain==0.1.0
pinecone-client==3.0.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2 
fastapi
uvicorn
python-dotenv
langchain
together
faiss-cpu
sentence-transformers
